fix: RGB565 color swap, task_id uint16 truncation, pipeline timeout
Three bugs fixed:

1. server/app/api/routes.py: RGB565 channel extraction was inverted.
   OV5640 FORMAT_CTRL00=0x6F (byte_swap=1) produces LE uint16 with
   bits[15:11]=Red, bits[10:5]=Green, bits[4:0]=Blue. Code had r/b
   swapped, rendering blue sky as yellow and vice-versa.

2. firmware/Core/Src/main.c: Three WiFi_HttpPostImage() call sites cast
   task_id to (uint16_t), silently truncating POSIX timestamps like
   1777153287 → 18439. The server's pipeline query looks for
   AnalysisResult.task_id >= original timestamp and never finds the
   truncated value. Removed the spurious casts — the function already
   takes uint32_t.

3. server/app/api/agent_routes.py: Pipeline timeout increased from
   max(15,...) to max(25,...) polls × 2s = 50s minimum, giving more
   headroom for 614KB RGB565 upload + Claude API analysis.

+++ firmware/Core/Src/main.c
@@ -1489,7 +1489,7 @@ static void _do_button_capture(void)
 
     /* Upload via HTTP POST */
     WiFiStatus_t wifi_ret = WiFi_HttpPostImage(
-        SERVER_UPLOAD_URL, (uint16_t)s_button_task_id,
+        SERVER_UPLOAD_URL, s_button_task_id,
         s_image_buffer, captured_size);
 
     uint32_t total_ms = HAL_GetTick() - perf_start;
@@ -1619,7 +1619,7 @@ static void _do_capture_now(void)
     /* Upload via HTTP POST (blocking, but _socket_send_all now
      * calls MQTT_ProcessLoop every 2s to keep the broker alive) */
     WiFiStatus_t wifi_ret = WiFi_HttpPostImage(
-        SERVER_UPLOAD_URL, (uint16_t)task_id,
+        SERVER_UPLOAD_URL, task_id,
         s_image_buffer, captured_size);
 
     uint32_t total_ms = HAL_GetTick() - perf_start;
@@ -1759,7 +1759,7 @@ static void _do_capture_sequence(void)
 
         /* Upload immediately */
         WiFiStatus_t wifi_ret = WiFi_HttpPostImage(
-            SERVER_UPLOAD_URL, (uint16_t)task_id,
+            SERVER_UPLOAD_URL, task_id,
             s_image_buffer, captured_size);
 
         snprintf(status_msg, sizeof(status_msg),
